--- .dockerignore
@@ -1,17 +1,0 @@
-**
-
-!src/**
-!public/**
-!next*
-!jest*
-!cypress/testData/**
-
-!sentry*
-!*.json
-!.npmrc
-!.babelrc
-!.sentryclirc
-
-!csp.js
-!prepare.js
-!triggBygg.md
