--- .github/workflows/deploy.yml
@@ -1,64 +1,0 @@
-name: Build and deploy
-
-on:
-  workflow_dispatch:
-  push:
-
-env:
-  IMAGE: ghcr.io/${{ github.repository }}/dp-produktside-frontend:${{ github.sha }}
-
-jobs:
-
-  build:
-    name: Build and publish Docker image
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@master
-
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          registry: ghcr.io
-          username: ${{ github.repository_owner }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Build and push
-        uses: docker/build-push-action@v2
-        with:
-          context: .
-          push: true
-          tags: ${{ env.IMAGE }}
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
-          secrets: |
-            "NODE_AUTH_TOKEN=${{ secrets.READER_TOKEN }}"
-            "SENTRY_AUTH_TOKEN=${{ secrets.SENTRY_AUTH_TOKEN }}"
-
-      - name: Archive NAIS yaml
-        uses: actions/upload-artifact@v1
-        with:
-          name: nais-yaml
-          path: nais
-
-  deploy-dev-gcp:
-    name: Deploy to dev gcp
-    needs: [build]
-    runs-on: ubuntu-latest
-    steps:
-      - name: Fetch NAIS yaml
-        uses: actions/download-artifact@v1
-        with:
-          name: nais-yaml
-          path: nais
-
-      - uses: nais/deploy/actions/deploy@v1
-        env:
-          APIKEY: ${{ secrets.NAIS_DEPLOY_APIKEY }}
-          CLUSTER: dev-gcp
-          RESOURCE: nais/nais.yaml
-          VARS: nais/dev-gcp/vars.yaml
-          PRINT_PAYLOAD: true
