--- nais/nais.yaml
@@ -1,40 +1,0 @@
-apiVersion: nais.io/v1alpha1
-kind: Application
-metadata:
-  name: dp-produktside-frontend
-  namespace: teamdagpenger
-  labels:
-    team: teamdagpenger
-spec:
-  image: {{ image }}
-  port: 3000
-  liveness:
-    path: /dagpenger/api/isalive
-  readiness:
-    path: /dagpenger/api/isready
-  prometheus:
-    enabled: false
-  replicas:
-    min: 2
-    max: 3
-  resources:
-    limits:
-      cpu: 500m
-      memory: 512Mi
-    requests:
-      cpu: 200m
-      memory: 256Mi
-  accessPolicy:
-    outbound:
-      external:
-        - host: www.nav.no
-        - host: dekoratoren.dev.nav.no
-  env:
-    - name: NODE_EXTRA_CA_CERTS
-      value: "/etc/pki/tls/certs/ca-bundle.crt"
-    - name: DEKORATOR_MILJO
-      value: {{dekoratorMiljo}}
-    - name: NPM_CONFIG_CACHE
-      value: /tmp
-  envFrom:
-  {{ingresses}}
