--- nais/prod-gcp/vars.yaml
@@ -1,3 +1,0 @@
-dekoratorMiljo: prod
-ingresses: |
-  ingresses:
