+++ next.config.js
@@ -1,14 +1,7 @@
-const path = require("path");
-
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
-  basePath: "/dagpenger",
-  productionBrowserSourceMaps: true,
-  sassOptions: {
-    includePaths: [path.join(__dirname, "src")],
-  },
-};
+}
 
-module.exports = nextConfig;
+module.exports = nextConfig
