+++ package.json
@@ -4,12 +4,11 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "prepare": "node ./prepare.js",
+    "prepare": "husky install",
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "check-updates": "ncu -i --format group"
+    "lint": "next lint"
   },
   "dependencies": {
     "@navikt/ds-css": "^0.18.27",
@@ -18,21 +17,21 @@
     "@navikt/nav-dekoratoren-moduler": "^1.6.17",
     "next": "12.2.3",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "4.7.4",
-    "@types/node": "18.6.3"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
+    "@types/node": "18.6.1",
     "@types/react": "18.0.15",
     "@types/react-dom": "18.0.6",
-    "eslint": "8.21.0",
+    "eslint": "8.20.0",
     "eslint-config-next": "12.2.3",
     "eslint-config-prettier": "^8.5.0",
     "husky": "^8.0.1",
     "lint-staged": "^13.0.3",
-    "npm-check-updates": "^16.0.4",
+    "npm-check-updates": "^16.0.0",
     "prettier": "^2.7.1",
-    "sass": "^1.54.0"
+    "sass": "^1.54.0",
+    "typescript": "4.7.4"
   },
   "prettier": {
     "printWidth": 120
