+++ pages/index.tsx
@@ -1,0 +1,26 @@
+import type { NextPage } from "next";
+import Head from "next/head";
+import Image from "next/image";
+import styles from "../styles/Home.module.scss";
+import Header from "../components/header/Header";
+
+const Home: NextPage = () => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className={styles.main}>
+        <Header />
+        <h2 className={styles.title}>
+          Welcome to <a href="https://nextjs.org">Next.js!</a>
+        </h2>
+      </main>
+    </div>
+  );
+};
+
+export default Home;
