--- prepare.js
@@ -1,5 +1,0 @@
-const isCi = process.env.CI !== undefined;
-
-if (!isCi) {
-  require("husky").install();
-}
