--- src/pages/api/isalive.js
@@ -1,4 +1,0 @@
-export default function handler(req, res) {
-  res.statusCode = 200;
-  res.end();
-}
