--- src/pages/api/isready.js
@@ -1,4 +1,0 @@
-export default async function handler(req, res) {
-  res.statusCode = 200;
-  res.end();
-}
